requirements.txt
selenium==4.15.2
requests==2.31.0
beautifulsoup4==4.12.3
openai==1.10.0
python-dotenv